\name{SSOSVM-package}
\alias{SSOSVM-package}
\alias{SSOSVM}
\docType{package}
\title{
\packageTitle{SSOSVM}
}
\description{
\packageDescription{SSOSVM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SSOSVM}
\packageIndices{SSOSVM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SSOSVM}

Maintainer: \packageMaintainer{SSOSVM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
