% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\docType{package}
\name{SSOSVM}
\alias{SSOSVM}
\alias{SSOSVM-package}
\title{SSOSVM: A package for online training of soft-margin support vector machines (SVMs) using the Stochastic majorization–minimization (SMM) algorithm.}
\description{
The SSOSVM package allows for computation...
}
\references{
Hien D. Nguyen, Andrew T. Jones and Geoffrey J. McLachlan. (2018). Stream-suitable optimization algorithms for some soft-margin support vector machine variants, Japanese Journal of Statistics and Data Science, vol. 1, Issue 1, pp. 81-108.
}
\author{
Andrew T. Jones,  Hien D. Nguyen, Geoffrey J. McLachlan
}
